fixed
添加测试面板，用户查看 CommandMatcher 的工作细节

+++ src/edu/sysu/pmglab/suranyi/commandParser/validator/RangeValidator.java
@@ -21,11 +21,11 @@ public class RangeValidator implements IValidator {
     }
 
     public RangeValidator(short MIN, short MAX) {
-        this((long) MIN, (long) MAX);
+        this(MIN, (long) MAX);
     }
 
     public RangeValidator(int MIN, int MAX) {
-        this((long) MIN, (long) MAX);
+        this(MIN, (long) MAX);
     }
 
     public RangeValidator(long MIN, long MAX) {
